chore: refactor UpgradeCheck to use DataSource
